requirements.txt
aiogram==3.1.0
apscheduler
python-dotenv
